+++ test/index.js
@@ -7,7 +7,6 @@ var
 var
   slice = [].slice;
 
-
 test('.getByKey()', function (assert) {
   var
     records = [
@@ -32,7 +31,6 @@ test('.getByKey()', function (assert) {
 
   assert.end();
 });
-
 
 test('.whitelist()', function (assert) {
   var
@@ -66,7 +64,6 @@ test('.whitelist()', function (assert) {
   assert.end();
 });
 
-
 test('.concat()', function (assert) {
   var
     records = [
@@ -92,745 +89,4 @@ test('.concat()', function (assert) {
     'should not alter original list.');
 
   assert.end();
-});
-
-
-test('.where() single clause', function (assert) {
-  // Find all orders by customer email
-  var records = [
-      {
-        "id": "ci6r6aliv00007poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "dennis@example.com",
-        "firstName": "Dennis",
-        "lastName": "Chambers",
-        "lineItems": {
-          "name": "Zildjian K Custom Organic Ride - 21\"",
-          "sku": "h617xrh",
-          "quantity": "1",
-          "total": "379.95"
-        }
-      },
-      {
-        "id": "ci6r6aliv00007poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "carlos@example.com",
-        "firstName": "Carlos",
-        "lastName": "Santana",
-        "lineItems": {
-          "name": "Gibson Memphis 1963 ES-335 TD - '60s Cherry 2014",
-          "sku": "sc37x3m",
-          "quantity": "1",
-          "total": "3999.00"
-        }
-      },
-      {
-        "id": "ci6r6aliv00007poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "dennis@example.com",
-        "firstName": "Dennis",
-        "lastName": "Chambers",
-        "lineItems": {
-          "name": "DW Collector's Series Metal Snare - 6.5\"x14\" Titanium 1mm",
-          "sku": "sc37x3m",
-          "quantity": "1",
-          "total": "379.95"
-        }
-      }
-    ],
-    copy = slice.call(records),
-
-    result = list(records).where({
-      billingEmail: 'dennis@example.com'
-    }),
-
-    expected = [
-      {
-        "id": "ci6r6aliv00007poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "dennis@example.com",
-        "firstName": "Dennis",
-        "lastName": "Chambers",
-        "lineItems": {
-          "name": "Zildjian K Custom Organic Ride - 21\"",
-          "sku": "h617xrh",
-          "quantity": "1",
-          "total": "379.95"
-        }
-      },
-      {
-        "id": "ci6r6aliv00007poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "dennis@example.com",
-        "firstName": "Dennis",
-        "lastName": "Chambers",
-        "lineItems": {
-          "name": "DW Collector's Series Metal Snare - 6.5\"x14\" Titanium 1mm",
-          "sku": "sc37x3m",
-          "quantity": "1",
-          "total": "379.95"
-        }
-      }
-    ];
-
-
-  assert.deepEqual(result, expected,
-    'should contain all the expected records.');
-
-  assert.equal(result.length, 2,
-    'should not contain excluded records.');
-
-  assert.deepEqual(records, copy,
-    'should not alter original list.');
-
-  assert.end();
-});
-
-
-test('.where() multiple clause', function (assert) {
-  // Find the order where Dennis ordered a cymbal.
-  var records = [
-      {
-        "id": "ci6r6aliv00007poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "dennis@example.com",
-        "firstName": "Dennis",
-        "lastName": "Chambers",
-        "lineItems": {
-          "name": "Zildjian K Custom Organic Ride - 21\"",
-          "sku": "h617xrh",
-          "quantity": "1",
-          "total": "379.95"
-        }
-      },
-      {
-        "id": "ci6r6aliv00007poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "carlos@example.com",
-        "firstName": "Carlos",
-        "lastName": "Santana",
-        "lineItems": {
-          "name": "Gibson Memphis 1963 ES-335 TD - '60s Cherry 2014",
-          "sku": "sc37x3m",
-          "quantity": "1",
-          "total": "3999.00"
-        }
-      },
-      {
-        "id": "ci6r6aliv00007poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "dennis@example.com",
-        "firstName": "Dennis",
-        "lastName": "Chambers",
-        "lineItems": {
-          "name": "DW Collector's Series Metal Snare - 6.5\"x14\" Titanium 1mm",
-          "sku": "sc37x3m",
-          "quantity": "1",
-          "total": "379.95"
-        }
-      }
-    ],
-    copy = slice.call(records),
-
-    result = list(records).where({
-      billingEmail: 'dennis@example.com',
-      lineItems: {
-        sku: 'h617xrh'
-      }
-    }),
-
-    expected = [
-      {
-        "id": "ci6r6aliv00007poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "dennis@example.com",
-        "firstName": "Dennis",
-        "lastName": "Chambers",
-        "lineItems": {
-          "name": "Zildjian K Custom Organic Ride - 21\"",
-          "sku": "h617xrh",
-          "quantity": "1",
-          "total": "379.95"
-        }
-      }
-    ];
-
-  assert.deepEqual(result, expected,
-    'should contain all the expected records.');
-
-  assert.equal(result.length, 1,
-    'should not contain excluded records.');
-
-  assert.deepEqual(records, copy,
-    'should not alter original list.');
-
-  assert.end();
-});
-
-test('.add() records to list', function (assert) {
-  var records = [
-      {
-        "id": "ci6r6aliv00007poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "dennis@example.com",
-        "firstName": "Dennis",
-        "lastName": "Chambers",
-        "lineItems": {
-          "name": "Zildjian K Custom Organic Ride - 21\"",
-          "sku": "h617xrh",
-          "quantity": "1",
-          "total": "379.95"
-        }
-      },
-      {
-        "id": "ci6r6aliv00007poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "carlos@example.com",
-        "firstName": "Carlos",
-        "lastName": "Santana",
-        "lineItems": {
-          "name": "Gibson Memphis 1963 ES-335 TD - '60s Cherry 2014",
-          "sku": "sc37x3m",
-          "quantity": "1",
-          "total": "3999.00"
-        }
-      },
-      {
-        "id": "ci6r6aliv00007poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "dennis@example.com",
-        "firstName": "Dennis",
-        "lastName": "Chambers",
-        "lineItems": {
-          "name": "DW Collector's Series Metal Snare - 6.5\"x14\" Titanium 1mm",
-          "sku": "sc37x3m",
-          "quantity": "1",
-          "total": "379.95"
-        }
-      }
-    ],
-    copy = slice.call(records),
-
-    record = {
-      id: 'ci6r6aliv00008poxc2zgnjvf',
-      date: '2014-12-30 05:29:28',
-      billingEmail: 'dennis@example.com',
-      firstName: 'Dennis',
-      lastName: 'Chambers',
-      lineItems: {
-        name: 'Zildjian K Custom Organic Ride - 21"',
-        sku: 'h123xrh',
-        quantity: '1',
-        total: '379.95'
-      }
-    },
-
-    result = list(records).add(record),
-
-    expected = result.where({ id: 'ci6r6aliv00008poxc2zgnjvf' })[0];
-
-  assert.deepEqual(record, expected,
-    'should contain the new record');
-
-  assert.deepEqual(records, copy,
-    'should not alter original list');
-
-  assert.end();
-});
-
-test('.add() async operation', function(assert) {
-  assert.plan(3);
-
-  var
-    records = [
-      {
-        "id": "ci6r6aliv00007poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "dennis@example.com",
-        "firstName": "Dennis",
-        "lastName": "Chambers",
-        "lineItems": {
-          "name": "Zildjian K Custom Organic Ride - 21\"",
-          "sku": "h617xrh",
-          "quantity": "1",
-          "total": "379.95"
-        }
-      }
-    ],
-    copy = slice.call(records),
-
-    newRecord = {
-      id: 'ci6r6aliv00008poxc2zgnjvf',
-      date: '2014-12-30 05:29:28',
-      billingEmail: 'dennis@example.com',
-      firstName: 'Dennis',
-      lastName: 'Chambers',
-      lineItems: {
-        name: 'Zildjian K Custom Organic Ride - 21"',
-        sku: 'h123xrh',
-        quantity: '1',
-        total: '379.95'
-      }
-    },
-
-    result = list({
-      list: records,
-      async: true
-    });
-
-  result = result.add(newRecord);
-  result.subscribe(
-    function onNext(newCollection) {
-      var expected = newCollection.where({ id: 'ci6r6aliv00008poxc2zgnjvf' })[0];
-
-      assert.deepEqual(newRecord, expected,
-        'should contain the new record');
-
-      assert.deepEqual(records, copy,
-        'should not alter original list');
-    },
-    function onError(err) {
-      assert.ok(!err, 'should not throw an error');
-    },
-    function onCompleted() {
-      assert.ok(true, 'should call onCompleted callback');
-
-      assert.end();
-    }
-  );
-});
-
-test('.remove() sync operation', function (assert) {
-  var records = [
-      {
-        "id": "ci6r6aliv00007poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "dennis@example.com",
-        "firstName": "Dennis",
-        "lastName": "Chambers",
-        "lineItems": {
-          "name": "Zildjian K Custom Organic Ride - 21\"",
-          "sku": "h617xrh",
-          "quantity": "1",
-          "total": "379.95"
-        }
-      },
-      {
-        "id": "ci6r6aliv00008poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "carlos@example.com",
-        "firstName": "Carlos",
-        "lastName": "Santana",
-        "lineItems": {
-          "name": "Gibson Memphis 1963 ES-335 TD - '60s Cherry 2014",
-          "sku": "sc37x3m",
-          "quantity": "1",
-          "total": "3999.00"
-        }
-      },
-      {
-        "id": "ci6r6aliv00007poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "dennis@example.com",
-        "firstName": "Dennis",
-        "lastName": "Chambers",
-        "lineItems": {
-          "name": "DW Collector's Series Metal Snare - 6.5\"x14\" Titanium 1mm",
-          "sku": "sc37x3m",
-          "quantity": "1",
-          "total": "379.95"
-        }
-      }
-    ],
-    copy = slice.call(records),
-
-    record = {
-      "id": "ci6r6aliv00008poxc2zgnjvf",
-      "date": "2014-12-30 05:29:28",
-      "billingEmail": "carlos@example.com",
-      "firstName": "Carlos",
-      "lastName": "Santana",
-      "lineItems": {
-        "name": "Gibson Memphis 1963 ES-335 TD - '60s Cherry 2014",
-        "sku": "sc37x3m",
-        "quantity": "1",
-        "total": "3999.00"
-      }
-    },
-
-    result = list(records).remove(record),
-
-    actual = result.where({ id: 'ci6r6aliv00008poxc2zgnjvf' })[0];
-
-  assert.deepEqual(actual, undefined,
-    'should not contain the removed record');
-
-  assert.deepEqual(records, copy,
-    'should not alter original list');
-
-  assert.end();
-});
-
-test('.remove() async operation', function(assert) {
-  assert.plan(3);
-
-  var
-    records = [
-      {
-        "id": "ci6r6aliv00007poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "dennis@example.com",
-        "firstName": "Dennis",
-        "lastName": "Chambers",
-        "lineItems": {
-          "name": "Zildjian K Custom Organic Ride - 21\"",
-          "sku": "h617xrh",
-          "quantity": "1",
-          "total": "379.95"
-        }
-      },
-      {
-        "id": "ci6r6aliv00008poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "carlos@example.com",
-        "firstName": "Carlos",
-        "lastName": "Santana",
-        "lineItems": {
-          "name": "Gibson Memphis 1963 ES-335 TD - '60s Cherry 2014",
-          "sku": "sc37x3m",
-          "quantity": "1",
-          "total": "3999.00"
-        }
-      },
-      {
-        "id": "ci6r6aliv00007poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "dennis@example.com",
-        "firstName": "Dennis",
-        "lastName": "Chambers",
-        "lineItems": {
-          "name": "DW Collector's Series Metal Snare - 6.5\"x14\" Titanium 1mm",
-          "sku": "sc37x3m",
-          "quantity": "1",
-          "total": "379.95"
-        }
-      }
-    ],
-    copy = slice.call(records),
-
-    record = {
-      "id": "ci6r6aliv00008poxc2zgnjvf",
-      "date": "2014-12-30 05:29:28",
-      "billingEmail": "carlos@example.com",
-      "firstName": "Carlos",
-      "lastName": "Santana",
-      "lineItems": {
-        "name": "Gibson Memphis 1963 ES-335 TD - '60s Cherry 2014",
-        "sku": "sc37x3m",
-        "quantity": "1",
-        "total": "3999.00"
-      }
-    },
-
-    result = list({
-      list: records,
-      async: true
-    });
-
-  result = result.remove(record);
-  result.subscribe(
-    function onNext(newCollection) {
-      var actual = newCollection.where({ id: 'ci6r6aliv00008poxc2zgnjvf' })[0];
-
-      assert.deepEqual(actual, undefined,
-        'should not contain the removed record');
-
-      assert.deepEqual(records, copy,
-        'should not alter original list');
-    },
-    function onError(err) {
-      assert.ok(!err, 'should not throw an error');
-    },
-    function onCompleted() {
-      assert.ok(true, 'should call onCompleted callback');
-
-      assert.end();
-    }
-  );
-});
-
-test('.removeWhere() sync operation', function (assert) {
-  var records = [
-      {
-        "id": "ci6r6aliv00007poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "dennis@example.com",
-        "firstName": "Dennis",
-        "lastName": "Chambers",
-        "lineItems": {
-          "name": "Zildjian K Custom Organic Ride - 21\"",
-          "sku": "h617xrh",
-          "quantity": "1",
-          "total": "379.95"
-        }
-      },
-      {
-        "id": "ci6r6aliv00008poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "carlos@example.com",
-        "firstName": "Carlos",
-        "lastName": "Santana",
-        "lineItems": {
-          "name": "Gibson Memphis 1963 ES-335 TD - '60s Cherry 2014",
-          "sku": "sc37x3m",
-          "quantity": "1",
-          "total": "3999.00"
-        }
-      },
-      {
-        "id": "ci6r6aliv00007poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "dennis@example.com",
-        "firstName": "Dennis",
-        "lastName": "Chambers",
-        "lineItems": {
-          "name": "DW Collector's Series Metal Snare - 6.5\"x14\" Titanium 1mm",
-          "sku": "sc37x3m",
-          "quantity": "1",
-          "total": "379.95"
-        }
-      }
-    ],
-    copy = slice.call(records),
-
-    query = {
-      "lastName": "Chambers",
-    },
-
-    result = list(records).removeWhere(query),
-
-    actual = result.where({ lastName: 'Chambers' })[0];
-
-  assert.deepEqual(actual, undefined,
-    'should not contain the removed record');
-
-  assert.deepEqual(records, copy,
-    'should not alter original list');
-
-  assert.end();
-});
-
-test('.removeWhere() async operation', function(assert) {
-  assert.plan(3);
-
-  var
-    records = [
-      {
-        "id": "ci6r6aliv00007poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "dennis@example.com",
-        "firstName": "Dennis",
-        "lastName": "Chambers",
-        "lineItems": {
-          "name": "Zildjian K Custom Organic Ride - 21\"",
-          "sku": "h617xrh",
-          "quantity": "1",
-          "total": "379.95"
-        }
-      },
-      {
-        "id": "ci6r6aliv00008poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "carlos@example.com",
-        "firstName": "Carlos",
-        "lastName": "Santana",
-        "lineItems": {
-          "name": "Gibson Memphis 1963 ES-335 TD - '60s Cherry 2014",
-          "sku": "sc37x3m",
-          "quantity": "1",
-          "total": "3999.00"
-        }
-      },
-      {
-        "id": "ci6r6aliv00007poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "dennis@example.com",
-        "firstName": "Dennis",
-        "lastName": "Chambers",
-        "lineItems": {
-          "name": "DW Collector's Series Metal Snare - 6.5\"x14\" Titanium 1mm",
-          "sku": "sc37x3m",
-          "quantity": "1",
-          "total": "379.95"
-        }
-      }
-    ],
-    copy = slice.call(records),
-
-    query = {
-      "lastName": "Chambers",
-    },
-
-    result = list({
-      list: records,
-      async: true
-    });
-
-  result = result.removeWhere(query);
-  result.subscribe(
-    function onNext(newCollection) {
-      var actual = newCollection.where({ lastName: 'Chambers' })[0];
-
-      assert.deepEqual(actual, undefined,
-        'should not contain the removed record');
-
-      assert.deepEqual(records, copy,
-        'should not alter original list');
-    },
-    function onError(err) {
-      assert.ok(!err, 'should not throw an error');
-    },
-    function onCompleted() {
-      assert.ok(true, 'should call onCompleted callback');
-
-      assert.end();
-    }
-  );
-});
-
-test('.removeSlice() sync operation', function (assert) {
-  var records = [
-      {
-        "id": "ci6r6aliv00007poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "dennis@example.com",
-        "firstName": "Dennis",
-        "lastName": "Chambers",
-        "lineItems": {
-          "name": "Zildjian K Custom Organic Ride - 21\"",
-          "sku": "h617xrh",
-          "quantity": "1",
-          "total": "379.95"
-        }
-      },
-      {
-        "id": "ci6r6aliv00008poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "carlos@example.com",
-        "firstName": "Carlos",
-        "lastName": "Santana",
-        "lineItems": {
-          "name": "Gibson Memphis 1963 ES-335 TD - '60s Cherry 2014",
-          "sku": "sc37x3m",
-          "quantity": "1",
-          "total": "3999.00"
-        }
-      },
-      {
-        "id": "ci6r6aliv00007poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "dennis@example.com",
-        "firstName": "Dennis",
-        "lastName": "Chambers",
-        "lineItems": {
-          "name": "DW Collector's Series Metal Snare - 6.5\"x14\" Titanium 1mm",
-          "sku": "sc37x3m",
-          "quantity": "1",
-          "total": "379.95"
-        }
-      }
-    ],
-    copy = slice.call(records),
-
-    result = list(records).removeSlice(1, 2),
-
-    actual = result.where({ id: 'ci6r6aliv00008poxc2zgnjvf' })[0];
-
-  assert.deepEqual(actual, undefined,
-    'should not contain the removed record');
-
-  assert.ok(result.length < copy.length,
-    'result should have less records than copy');
-
-  assert.deepEqual(records, copy,
-    'should not alter original list');
-
-  assert.end();
-});
-
-test('.removeSlice() async operation', function(assert) {
-  assert.plan(4);
-
-  var
-    records = [
-      {
-        "id": "ci6r6aliv00007poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "dennis@example.com",
-        "firstName": "Dennis",
-        "lastName": "Chambers",
-        "lineItems": {
-          "name": "Zildjian K Custom Organic Ride - 21\"",
-          "sku": "h617xrh",
-          "quantity": "1",
-          "total": "379.95"
-        }
-      },
-      {
-        "id": "ci6r6aliv00008poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "carlos@example.com",
-        "firstName": "Carlos",
-        "lastName": "Santana",
-        "lineItems": {
-          "name": "Gibson Memphis 1963 ES-335 TD - '60s Cherry 2014",
-          "sku": "sc37x3m",
-          "quantity": "1",
-          "total": "3999.00"
-        }
-      },
-      {
-        "id": "ci6r6aliv00007poxc2zgnjvf",
-        "date": "2014-12-30 05:29:28",
-        "billingEmail": "dennis@example.com",
-        "firstName": "Dennis",
-        "lastName": "Chambers",
-        "lineItems": {
-          "name": "DW Collector's Series Metal Snare - 6.5\"x14\" Titanium 1mm",
-          "sku": "sc37x3m",
-          "quantity": "1",
-          "total": "379.95"
-        }
-      }
-    ],
-    copy = slice.call(records),
-
-    result = list({
-      list: records,
-      async: true
-    });
-
-  result = result.removeSlice(1, 2);
-  result.subscribe(
-    function onNext(newCollection) {
-      var actual = newCollection.where({ id: 'ci6r6aliv00008poxc2zgnjvf' })[0];
-
-      assert.deepEqual(actual, undefined,
-        'should not contain the removed record');
-
-      assert.ok(newCollection.length < copy.length,
-        'newCollection should have less records than copy');
-
-      assert.deepEqual(records, copy,
-        'should not alter original list');
-    },
-    function onError(err) {
-      assert.ok(!err, 'should not throw an error');
-    },
-    function onCompleted() {
-      assert.ok(true, 'should call onCompleted callback');
-
-      assert.end();
-    }
-  );
 });
